+++ mllib/src/main/scala/org/apache/spark/ml/tree/Node.scala
@@ -136,7 +136,8 @@ final class LeafNode private[ml] (
 
   override private[ml] def toOld(id: Int): OldNode = {
     new OldNode(id, new OldPredict(prediction, prob = impurityStats.prob(prediction)),
-      impurity, isLeaf = true, None, None, None, None)
+      impurity, isLeaf = true, None, None, None,
+      Some(OldInformationGainStats.invalidInformationGainStats))
   }
 
   override private[ml] def maxSplitFeatureIndex(): Int = -1
